+++ DearPyGui/src/Core/PlotAppItems/mvLineSeries.h
@@ -27,7 +27,7 @@ namespace Marvel {
 
 			ImPlot::SetNextLineStyle(m_color.toVec4());
 			ImPlot::PushStyleVar(ImPlotStyleVar_LineWeight, m_weight);
-			ImPlot::PlotLine(m_name.c_str(), m_xs.data(), m_ys.data(), m_xs.size());
+			ImPlot::PlotLine(m_name.c_str(), m_xs.data(), m_ys.data(), (int)m_xs.size());
 			ImPlot::PopStyleVar();
 		}
